rag-gym-impl/pyproject.toml
[project]
name = "rag-gym-impl"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "python-dotenv>=1.0.0",
    "ragflow-sdk>=0.1.0",
]

[tool.poetry]
packages = [
    {include = "rag_gym_impl", from = "src"},
]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
debugpy = "<2.0.0"

[tool.pytest.ini_options]
markers = [
    'integration: marks tests that call external services (deselect with "-m "not integration"")',
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
